refactor(preview-url): pull the URL contract out of the electron-bound module
decision(preview-url): scheme 名 / preview id の契約 / URL の組み立て・分解を electron 非依存の
モジュールへ移す。配信の実装は protocol と fs に依存するが、防壁の判定と RPC 入口の検証は
純関数として bun test から呼ぶ。同居させると test 側が electron の import で落ちる
learned(url): host は URL パーサが正規化する。配信許可の登録キーは生の id なので、正規化で
変わる文字が混じると「登録は成功するのに配信だけ 403」になり、ログには out-of-root としか
出ず原因が見えない。id の形式を入口で固定して生成元を変えても壊れないようにする
